boiler: move dummy db impl to internal pkg
