+++ cmd/stoker/main.go
@@ -10,6 +10,10 @@ import (
 	"github.com/frantjc/sindri/command"
 	xerrors "github.com/frantjc/x/errors"
 	xos "github.com/frantjc/x/os"
+	_ "github.com/moby/buildkit/client/connhelper/dockercontainer"
+	_ "github.com/moby/buildkit/client/connhelper/kubepod"
+	_ "github.com/moby/buildkit/client/connhelper/nerdctlcontainer"
+	_ "github.com/moby/buildkit/client/connhelper/podmancontainer"
 )
 
 func main() {
